+++ uv.gyp
@@ -237,6 +237,7 @@
             'src/unix/linux-inotify.c',
             'src/unix/linux-syscalls.c',
             'src/unix/linux-syscalls.h',
+            'src/unix/sysinfo-loadavg.c',
             'src/unix/sysinfo-memory.c',
           ],
           'link_settings': {
@@ -252,6 +253,7 @@
             'src/unix/pthread-fixes.c',
             'src/unix/android-ifaddrs.c',
             'src/unix/pthread-barrier.c'
+            'src/unix/sysinfo-loadavg.c',
             'src/unix/sysinfo-memory.c',
           ],
           'link_settings': {
